requirements.txt
beautifulsoup4==4.4.1
decorator==4.0.9
praw==3.5.0
praw-oauth2util==0.3.4
requests==2.10.0
six==1.10.0
update-checker==0.11
psycopg2==2.6.1
urlparse2==1.1.1